+++ examples/29_dsn_parsing.py
@@ -21,7 +21,6 @@ print('Multiple ranges with multiple ports and with default port at the end: ')
 result = C._process_dsn('127.0.0.10..19:8564,127.0.0.20,localhost:8565,127.0.0.21..23')
 printer.pprint(sorted(result))
 
-
 # Empty DSN
 try:
     result = C._process_dsn(' ')
@@ -39,7 +38,14 @@ except pyexasol.ExaConnectionDsnError as e:
 
 # Cannot resolve hostname
 try:
-    result = C._process_dsn('testl1..5.zlan')
+    result = C._process_dsn('test1..5.zlan')
+    printer.pprint(result)
+except pyexasol.ExaConnectionDsnError as e:
+    print(e)
+
+# Hostname range with zero-padding
+try:
+    result = C._process_dsn('test01..20.zlan')
     printer.pprint(result)
 except pyexasol.ExaConnectionDsnError as e:
     print(e)