+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### ExaConnection
 
+- DSN hostname ranges with zero-padding are now supported (e.g. `myhost01..16`).
 - Context manager ("with" statement) is now supported for connection object.
 
 ### ExaStatement